+++ learning_projects/MovieAppCleanArchitecture/app/src/main/java/com/example/movieappcleanarchitecture/presentation/screens/dashboard/DashboardViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.movieappcleanarchitecture.presentation.viewmodels
+package com.example.movieappcleanarchitecture.presentation.screens.dashboard
 
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
